# openapi/openapi_internal_massive.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "PagoPA API Debt Position",
    "description": "Progetto Gestione Posizioni Debitorie",
    "termsOfService": "https://www.pagopa.gov.it/",
    "version": "0.11.82"
  },
  "servers": [
    {
      "url": "https://api.uat.platform.pagopa.it/gpd/debt-positions-service/v1/",
      "description": "GPD Test environment"
    },
    {
      "url": "https://api.platform.pagopa.it/gpd/debt-positions-service/v1/",
      "description": "GPD Production Environment"
    },
    {
      "url": "https://api.uat.platform.pagopa.it/gpd/debt-positions-service/v3/",
      "description": "GPD Test environment"
    },
    {
      "url": "https://api.platform.pagopa.it/gpd/debt-positions-service/v3/",
      "description": "GPD Production Environment"
    },
    {
      "url": "https://api.uat.platform.pagopa.it/aca/debt-positions-service/v1/",
      "description": "ACA Test environment"
    },
    {
      "url": "https://api.platform.pagopa.it/aca/debt-positions-service/v1/",
      "description": "ACA Production environment"
    }
  ],
...
  "paths": {
...
    "/organizations/{organizationfiscalcode}/debtpositions/{iupd}/invalidate": {
      "post": {
        "tags": [
          "Debt Position Actions API"
        ],
        "summary": "The Organization invalidate a debt Position.",
        "operationId": "invalidatePosition",
        "parameters": [
          {
            "name": "organizationfiscalcode",
            "in": "path",
            "description": "Organization fiscal code, the fiscal code of the Organization.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "iupd",
            "in": "path",
            "description": "IUPD (Unique identifier of the debt position). Format could be `<Organization fiscal code + UUID>` this would make it unique within the new PD management system. It's the responsibility of the EC to guarantee uniqueness. The pagoPa system shall verify that this is `true` and if not, notify the EC.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "404": {
            "description": "No debt position found.",
            "headers": {
              "X-Request-Id": {
                "description": "This header identifies the call",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemJson"
                }
              }
            }
          },
          "409": {
            "description": "Conflict: debt position is not in invalidable state.",
            "headers": {
              "X-Request-Id": {
                "description": "This header identifies the call",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemJson"
                }
              }
            }
          },
          "200": {
            "description": "Request published.",
            "headers": {
              "X-Request-Id": {
                "description": "This header identifies the call",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentPositionModel"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "headers": {
              "X-Request-Id": {
                "description": "This header identifies the call",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "examples": {
                  "forbidden": {
                    "description": "forbidden",
                    "value": {
                      "statusCode": 403,
                      "message": "You are not allowed to access this resource."
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Service unavailable.",
            "headers": {
              "X-Request-Id": {
                "description": "This header identifies the call",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemJson"
                }
              }
            }
          },
          "401": {
            "description": "Wrong or missing function key.",
            "headers": {
              "X-Request-Id": {
                "description": "This header identifies the call",
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "security": [
          {
            "ApiKey": []
          },
          {
            "Authorization": []
          }
        ]
      },
      "parameters": [
        {
          "name": "X-Request-Id",
          "in": "header",
          "description": "This header identifies the call, if not passed it is self-generated. This ID is returned in the response.",
          "schema": {
            "type": "string"
          }
        }
      ]
    },
    "/organizations/{organizationfiscalcode}/debtpositions/{iupd}/publish": {
      "post": {
        "tags": [
          "Debt Position Actions API"
        ],
        "summary": "The Organization publish a debt Position.",
        "operationId": "publishPosition_1",
        "parameters": [
          {
            "name": "organizationfiscalcode",
            "in": "path",
            "description": "Organization fiscal code, the fiscal code of the Organization.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "iupd",
            "in": "path",
            "description": "IUPD (Unique identifier of the debt position). Format could be `<Organization fiscal code + UUID>` this would make it unique within the new PD management system. It's the responsibility of the EC to guarantee uniqueness. The pagoPa system shall verify that this is `true` and if not, notify the EC.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "404": {
            "description": "No debt position found.",
            "headers": {
              "X-Request-Id": {
                "description": "This header identifies the call",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemJson"
                }
              }
            }
          },
          "409": {
            "description": "Conflict: debt position is not in publishable state.",
            "headers": {
              "X-Request-Id": {
                "description": "This header identifies the call",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemJson"
                }
              }
            }
          },
          "200": {
            "description": "Request published.",
            "headers": {
              "X-Request-Id": {
                "description": "This header identifies the call",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentPositionModel"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "headers": {
              "X-Request-Id": {
                "description": "This header identifies the call",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "examples": {
                  "forbidden": {
                    "description": "forbidden",
                    "value": {
                      "statusCode": 403,
                      "message": "You are not allowed to access this resource."
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Service unavailable.",
            "headers": {
              "X-Request-Id": {
                "description": "This header identifies the call",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemJson"
                }
              }
            }
          },
          "401": {
            "description": "Wrong or missing function key.",
            "headers": {
              "X-Request-Id": {
                "description": "This header identifies the call",
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "security": [
          {
            "ApiKey": []
          },
          {
            "Authorization": []
          }
        ]
      },
      "parameters": [
        {
          "name": "X-Request-Id",
          "in": "header",
          "description": "This header identifies the call, if not passed it is self-generated. This ID is returned in the response.",
          "schema": {
            "type": "string"
          }
        }
      ]
    },
...
    "/v3/organizations/{organizationfiscalcode}/debtpositions/{iupd}/publish": {
      "post": {
        "tags": [
          "Debt Position Actions API"
        ],
        "summary": "The Organization publish a debt Position.",
        "operationId": "publishPosition",
        "parameters": [
          {
            "name": "organizationfiscalcode",
            "in": "path",
            "description": "Organization fiscal code, the fiscal code of the Organization.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "iupd",
            "in": "path",
            "description": "IUPD (Unique identifier of the debt position). Format could be `<Organization fiscal code + UUID>` this would make it unique within the new PD management system. It's the responsibility of the EC to guarantee uniqueness. The pagoPa system shall verify that this is `true` and if not, notify the EC.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "404": {
            "description": "No debt position found.",
            "headers": {
              "X-Request-Id": {
                "description": "This header identifies the call",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemJson"
                }
              }
            }
          },
          "409": {
            "description": "Conflict: debt position is not in publishable state.",
            "headers": {
              "X-Request-Id": {
                "description": "This header identifies the call",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemJson"
                }
              }
            }
          },
          "200": {
            "description": "Request published.",
            "headers": {
              "X-Request-Id": {
                "description": "This header identifies the call",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentPositionModelV3"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "headers": {
              "X-Request-Id": {
                "description": "This header identifies the call",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "examples": {
                  "forbidden": {
                    "description": "forbidden",
                    "value": {
                      "statusCode": 403,
                      "message": "You are not allowed to access this resource."
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Service unavailable.",
            "headers": {
              "X-Request-Id": {
                "description": "This header identifies the call",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemJson"
                }
              }
            }
          },
          "401": {
            "description": "Wrong or missing function key.",
            "headers": {
              "X-Request-Id": {
                "description": "This header identifies the call",
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "security": [
          {
            "ApiKey": []
          }
        ]
      },
      "parameters": [
        {
          "name": "X-Request-Id",
          "in": "header",
          "description": "This header identifies the call, if not passed it is self-generated. This ID is returned in the response.",
          "schema": {
            "type": "string"
          }
        }
      ]
    }
  },
  "components": {
    "schemas": {
      "DebtorModel": {
        "required": [
          "fiscalCode",
          "fullName",
          "type"
        ],
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "F",
              "G"
            ]
          },
          "fiscalCode": {
            "type": "string"
          },
          "fullName": {
            "type": "string"
          },
          "streetName": {
            "type": "string"
          },
          "civicNumber": {
            "type": "string"
          },
          "postalCode": {
            "type": "string"
          },
          "city": {
            "type": "string"
          },
          "province": {
            "type": "string"
          },
          "region": {
            "type": "string"
          },
          "country": {
            "pattern": "[A-Z]{2}",
            "type": "string",
            "example": "IT"
          },
          "email": {
            "type": "string",
            "example": "email@domain.com"
          },
          "phone": {
            "type": "string"
          }
        }
      },
...
      "InstallmentModel": {
        "required": [
          "amount",
          "description",
          "dueDate",
          "iuv"
        ],
        "type": "object",
        "properties": {
          "nav": {
            "type": "string"
          },
          "iuv": {
            "type": "string"
          },
          "amount": {
            "type": "integer",
            "format": "int64"
          },
          "description": {
            "maxLength": 140,
            "minLength": 0,
            "type": "string"
          },
          "dueDate": {
            "type": "string",
            "format": "date-time"
          },
          "fee": {
            "type": "integer",
            "format": "int64",
            "readOnly": true
          },
          "notificationFee": {
            "type": "integer",
            "format": "int64",
            "readOnly": true
          },
          "status": {
            "type": "string",
            "readOnly": true,
            "enum": [
              "UNPAID",
              "PAID",
              "PARTIALLY_REPORTED",
              "REPORTED",
              "UNPAYABLE",
              "EXPIRED"
            ]
          },
          "transfer": {
            "maxItems": 2147483647,
            "minItems": 1,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TransferModel"
            }
          },
          "installmentMetadata": {
            "maxItems": 10,
            "minItems": 0,
            "type": "array",
            "description": "it can added a maximum of 10 key-value pairs for metadata",
            "items": {
              "$ref": "#/components/schemas/InstallmentMetadataModel"
            }
          }
        }
      },
      "PaymentOptionModelV3": {
        "required": [
          "debtor",
          "installments",
          "switchToExpired"
        ],
        "type": "object",
        "properties": {
          "description": {
            "maxLength": 140,
            "minLength": 0,
            "type": "string",
            "writeOnly": true
          },
          "validityDate": {
            "type": "string",
            "format": "date-time"
          },
          "retentionDate": {
            "type": "string",
            "format": "date-time"
          },
          "switchToExpired": {
            "type": "boolean",
            "description": "feature flag to enable the payment option to expire after the due date",
            "example": false,
            "default": false
          },
          "debtor": {
            "$ref": "#/components/schemas/DebtorModel"
          },
          "installments": {
            "maxItems": 100,
            "minItems": 1,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/InstallmentModel"
            }
          }
        }
      },
      "PaymentPositionModelV3": {
        "required": [
          "companyName",
          "iupd",
          "paymentOption"
        ],
        "type": "object",
        "properties": {
          "iupd": {
            "type": "string"
          },
          "payStandIn": {
            "type": "boolean",
            "description": "feature flag to enable a debt position in stand-in mode",
            "example": true,
            "default": true
          },
          "companyName": {
            "maxLength": 140,
            "minLength": 0,
            "type": "string"
          },
          "officeName": {
            "maxLength": 140,
            "minLength": 0,
            "type": "string"
          },
          "paymentDate": {
            "type": "string",
            "format": "date-time",
            "readOnly": true
          },
          "status": {
            "type": "string",
            "readOnly": true,
            "enum": [
              "DRAFT",
              "PUBLISHED",
              "VALID",
              "UNPAYABLE",
              "PARTIALLY_PAID",
              "PAID"
            ]
          },
          "paymentOption": {
            "maxItems": 100,
            "minItems": 1,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PaymentOptionModelV3"
            }
          }
        }
      },
...
      "TransferModel": {
        "required": [
          "amount",
          "category",
          "idTransfer",
          "remittanceInformation"
        ],
        "type": "object",
        "properties": {
          "idTransfer": {
            "type": "string",
            "enum": [
              "1",
              "2",
              "3",
              "4",
              "5"
            ]
          },
          "amount": {
            "type": "integer",
            "format": "int64"
          },
          "organizationFiscalCode": {
            "type": "string",
            "description": "Fiscal code related to the organization targeted by this transfer.",
            "example": "00000000000"
          },
          "remittanceInformation": {
            "maxLength": 140,
            "minLength": 0,
            "type": "string"
          },
          "category": {
            "type": "string"
          },
          "iban": {
            "type": "string",
            "description": "mutual exclusive with stamp",
            "example": "IT0000000000000000000000000"
          },
          "postalIban": {
            "type": "string",
            "description": "optional - can be combined with iban but not with stamp",
            "example": "IT0000000000000000000000000"
          },
          "stamp": {
            "$ref": "#/components/schemas/Stamp"
          },
          "companyName": {
            "maxLength": 140,
            "minLength": 0,
            "type": "string"
          },
          "transferMetadata": {
            "maxItems": 10,
            "minItems": 0,
            "type": "array",
            "description": "it can added a maximum of 10 key-value pairs for metadata",
            "items": {
              "$ref": "#/components/schemas/TransferMetadataModel"
            }
          }
        }
      },
      "ProblemJson": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "description": "A short, summary of the problem type. Written in english and readable for engineers (usually not suited for non technical stakeholders and not localized); example: Service Unavailable"
          },
          "status": {
            "maximum": 600,
            "minimum": 100,
            "type": "integer",
            "description": "The HTTP status code generated by the origin server for this occurrence of the problem.",
            "format": "int32",
            "example": 200
          },
          "detail": {
            "type": "string",
            "description": "A human readable explanation specific to this occurrence of the problem.",
            "example": "There was an error processing the request"
          }
        }
      },
...
      "PaymentOptionMetadataModel": {
        "required": [
          "key"
        ],
        "type": "object",
        "properties": {
          "key": {
            "type": "string"
          },
          "value": {
            "type": "string"
          }
        },
        "description": "it can added a maximum of 10 key-value pairs for metadata"
      },
      "PaymentOptionModel": {
        "required": [
          "amount",
          "description",
          "dueDate",
          "isPartialPayment",
          "iuv"
        ],
        "type": "object",
        "properties": {
          "nav": {
            "type": "string"
          },
          "iuv": {
            "type": "string"
          },
          "amount": {
            "type": "integer",
            "format": "int64"
          },
          "description": {
            "maxLength": 140,
            "minLength": 0,
            "type": "string"
          },
          "isPartialPayment": {
            "type": "boolean"
          },
          "dueDate": {
            "type": "string",
            "format": "date-time"
          },
          "retentionDate": {
            "type": "string",
            "format": "date-time"
          },
          "fee": {
            "type": "integer",
            "format": "int64"
          },
          "notificationFee": {
            "type": "integer",
            "format": "int64",
            "readOnly": true
          },
          "transfer": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TransferModel"
            }
          },
          "paymentOptionMetadata": {
            "maxItems": 10,
            "minItems": 0,
            "type": "array",
            "description": "it can added a maximum of 10 key-value pairs for metadata",
            "items": {
              "$ref": "#/components/schemas/PaymentOptionMetadataModel"
            }
          }
        }
      },
      "PaymentPositionModel": {
        "required": [
          "companyName",
          "fiscalCode",
          "fullName",
          "iupd",
          "switchToExpired",
          "type"
        ],
        "type": "object",
        "properties": {
          "iupd": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "enum": [
              "F",
              "G"
            ]
          },
          "payStandIn": {
            "type": "boolean",
            "description": "feature flag to enable a debt position in stand-in mode",
            "example": true,
            "default": true
          },
          "fiscalCode": {
            "type": "string"
          },
          "fullName": {
            "type": "string"
          },
          "streetName": {
            "type": "string"
          },
          "civicNumber": {
            "type": "string"
          },
          "postalCode": {
            "type": "string"
          },
          "city": {
            "type": "string"
          },
          "province": {
            "type": "string"
          },
          "region": {
            "type": "string"
          },
          "country": {
            "pattern": "[A-Z]{2}",
            "type": "string",
            "example": "IT"
          },
          "email": {
            "type": "string",
            "example": "email@domain.com"
          },
          "phone": {
            "type": "string"
          },
          "switchToExpired": {
            "type": "boolean",
            "description": "feature flag to enable the debt position to expire after the due date",
            "example": false,
            "default": false
          },
          "companyName": {
            "maxLength": 140,
            "minLength": 0,
            "type": "string"
          },
          "officeName": {
            "maxLength": 140,
            "minLength": 0,
            "type": "string"
          },
          "validityDate": {
            "type": "string",
            "format": "date-time"
          },
          "paymentDate": {
            "type": "string",
            "format": "date-time",
            "readOnly": true
          },
          "status": {
            "type": "string",
            "readOnly": true,
            "enum": [
              "DRAFT",
              "PUBLISHED",
              "VALID",
              "INVALID",
              "EXPIRED",
              "PARTIALLY_PAID",
              "PAID",
              "REPORTED"
            ]
          },
          "paymentOption": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PaymentOptionModel"
            }
          }
        }
      },
...
    },
    "securitySchemes": {
      "ApiKey": {
        "type": "apiKey",
        "description": "The API key to access this function app.",
        "name": "Ocp-Apim-Subscription-Key",
        "in": "header"
      },
      "Authorization": {
        "type": "http",
        "description": "JWT token get after Azure Login",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}